+++ packages/core/src/interfaces/variants.ts
@@ -13,11 +13,6 @@ export interface MediaQuery {
 	readonly value: string;
 }
 
-export interface ContainerQuery {
-	readonly prefix: string;
-	readonly value: string;
-}
-
 export interface Opacity {
 	readonly prefix: string;
 	readonly value: string;
@@ -27,6 +22,5 @@ export interface Variants {
 	readonly pseudoClasses?: readonly PseudoClass[];
 	readonly pseudoElements?: readonly PseudoElement[];
 	readonly mediaQueries?: readonly MediaQuery[];
-	readonly containerQueries?: readonly ContainerQuery[];
 	readonly opacity?: readonly Opacity[];
 }